requirements.txt
Pillow~=8.2.0
numpy~=1.19.5
matplotlib~=3.4.1
boto3~=1.17.47
ipython~=7.22.0
requests~=2.25.1
flask~=1.1.2
tensorflow~=2.4.1
Flask-Cors~=3.0.10
opencv-python~=4.5.1.48
tensorflow-hub~=0.12.0